+++ shared-libs/crates/start-core/src/db/model/public.rs
@@ -74,7 +74,7 @@ impl Public {
                                         add_ssl: Some(AddSslOptions {
                                             preferred_external_port: 443,
                                             add_x_forwarded_headers: false,
-                                            alpn: Some(AlpnInfo::Specified(vec![
+                                            alpn: Some(AlpnInfo(vec![
                                                 MaybeUtf8String("h2".into()),
                                                 MaybeUtf8String("http/1.1".into()),
                                             ])),